Add LICENSE
The code is licensed as CC0.

The data is licensed under various CC licenses and will be part of a
separate repo. Importantly, there is data that may not be modified.

Where possible I will upload processed data as a release on Github.
Users will have to locally process the remaining data.