Add provenance information
This enables a more precise merging of frequencies from two lists.
If the term is counted differently in the same literary source then
we can clearly take the maximum from both counts as the total count.
Counts from different sources are distinct and counted as such.

Adds support for provenance to CHJ premodern.